requirements.txt
# T-Mobile Installation Cost Prediction API Dependencies
# Optimized for Windows compatibility with pre-compiled packages

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Configuration management
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ML and data processing (using pre-compiled versions)
pandas>=2.0.0,<3.0.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
xgboost>=1.7.0,<3.0.0
joblib>=1.3.0

# HTTP client for testing
requests>=2.28.0

# Future vault integration (optional)
# hvac==2.0.0

# Development and testing (optional - install separately if needed)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2
